Included middleware and nodemon for hot reloading; Inclusions: cookie-parser; compression; morgan; cors; note: Helmet is advised for improved security but not yet included
